+++ src/api_server/src/parsed_request.rs
@@ -27,6 +27,7 @@ use logger::{error, info};
 use vmm::rpc_interface::{VmmAction, VmmActionError};
 
 pub(crate) enum ParsedRequest {
+    GetMmdsVersion,
     GetMMDS,
     PatchMMDS(Value),
     PutMMDS(Value),
@@ -63,7 +64,7 @@ impl ParsedRequest {
                 Ok(ParsedRequest::new_sync(VmmAction::GetFullVmConfig))
             }
             (Method::Get, "machine-config", None) => parse_get_machine_config(),
-            (Method::Get, "mmds", None) => parse_get_mmds(),
+            (Method::Get, "mmds", None) => parse_get_mmds(path_tokens.get(1)),
             (Method::Get, _, Some(_)) => method_to_error(Method::Get),
             (Method::Put, "actions", Some(body)) => parse_put_actions(body),
             (Method::Put, "balloon", Some(body)) => parse_put_balloon(body),
@@ -280,6 +281,7 @@ pub(crate) mod tests {
                 (&ParsedRequest::Sync(ref sync_req), &ParsedRequest::Sync(ref other_sync_req)) => {
                     sync_req == other_sync_req
                 }
+                (&ParsedRequest::GetMmdsVersion, &ParsedRequest::GetMmdsVersion) => true,
                 (&ParsedRequest::GetMMDS, &ParsedRequest::GetMMDS) => true,
                 (&ParsedRequest::PutMMDS(ref val), &ParsedRequest::PutMMDS(ref other_val)) => {
                     val == other_val
@@ -596,6 +598,13 @@ pub(crate) mod tests {
         let mut connection = HttpConnection::new(receiver);
         sender
             .write_all(http_request("GET", "/mmds", None).as_bytes())
+            .unwrap();
+        assert!(connection.try_read().is_ok());
+        let req = connection.pop_parsed_request().unwrap();
+        assert!(ParsedRequest::try_from_request(&req).is_ok());
+
+        sender
+            .write_all(http_request("GET", "/mmds/version", None).as_bytes())
             .unwrap();
         assert!(connection.try_read().is_ok());
         let req = connection.pop_parsed_request().unwrap();